+++ tests/float_test.rs
@@ -5,7 +5,7 @@ use winnow_grammar::grammar;
 grammar! {
     grammar FloatParser {
         pub rule test_float -> f64 =
-            f:float -> { f }
+            f:f64 -> { f }
     }
 }
